move useRenameShare()
