+++ projects/graphql-client/src/index.ts
@@ -25,3 +25,4 @@ export * from "./remote/mutations/useUpdateShareMemberPermissions"
 export * from "./remote/mutations/useUpdatePlaylistSongOrder"
 export * from "./remote/mutations/useRevokeInvitation"
 export * from "./remote/mutations/useInviteToShare"
+export * from "./remote/mutations/useRenameShare"
